requirements.txt
altgraph==0.17.4
build==1.2.2.post1
cffi==1.17.1
colorama==0.4.6
cx-Freeze==6.15.16
cx-Logging==3.2.0
importlib-metadata==8.5.0
lief==0.14.1
packaging==25.0
pefile==2023.2.7
pillow==10.4.0
pycparser==2.22
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
pyproject-hooks==1.2.0
pyvips==3.0.0
pywin32-ctypes==0.2.3
tomli==2.2.1
zipp==3.20.2
